added relationships between task and user + refactored

+++ src/tasks/tasks.service.ts
@@ -1,10 +1,11 @@
-import {Injectable, NotFoundException} from '@nestjs/common';
-import {InjectRepository} from '@nestjs/typeorm';
-import {CreateTaskDto} from './dto/create-task.dto';
-import {GetTaskFilterDto} from './dto/get-task-filter.dto';
-import {TaskStatus} from './task-status.enum';
-import {Task} from './task.entity';
-import {TaskRepository} from './task.repository';
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
+import { CreateTaskDto } from './dto/create-task.dto';
+import { GetTaskFilterDto } from './dto/get-task-filter.dto';
+import { TaskStatus } from './task-status.enum';
+import { Task } from './task.entity';
+import { TaskRepository } from './task.repository';
 
 /**
  * Services serves the business logic
@@ -13,39 +14,45 @@ import {TaskRepository} from './task.repository';
 
 @Injectable()
 export class TasksService {
-    constructor(
-        @InjectRepository(TaskRepository)
-        private taskRepository: TaskRepository,
-    ) {}
+  constructor(
+    @InjectRepository(TaskRepository)
+    private taskRepository: TaskRepository,
+  ) {}
 
-    async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-        return this.taskRepository.getTasks(filterDto);
-    }
+  async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
+  }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
-        if (!found) {
-            throw new NotFoundException(`Task with ${id} not found`);
-        }
-        return found;
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
+    if (!found) {
+      throw new NotFoundException(`Task with ${id} not found`);
     }
+    return found;
+  }
 
-    async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-        return this.taskRepository.createTask(createTaskDto);
-    }
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
+  }
 
-    async deleteTask(id: number): Promise<void> {
-        // contains number of item deleted
-        const result = await this.taskRepository.delete(id);
-        if (result.affected === 0) {
-            throw new NotFoundException(`Task with ${id} not found`);
-        }
+  async deleteTask(id: number, user): Promise<void> {
+    // contains number of item deleted
+    const result = await this.taskRepository.delete({ id, userId: user.id });
+    if (result.affected === 0) {
+      throw new NotFoundException(`Task with ${id} not found`);
     }
+  }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-        const task = await this.getTaskById(id);
-        task.status = status;
-        await task.save();
-        return task;
-    }
+  async updateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskById(id, user);
+    task.status = status;
+    await task.save();
+    return task;
+  }
 }
